Redux enabled, Context/Reducer disabled
