+++ envolve-frontend/src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import './App.css';
 import {createMuiTheme, MuiThemeProvider} from "@material-ui/core";
-import UserContextProvider from "./context/UserContextProvider";
 import Navigation from "./components/Navigation";
 
 
@@ -15,7 +14,7 @@ const theme = createMuiTheme({
         }
     },
     typography: {
-        fontFamily: ['Josefin Sans', 'sans-serif', 'Modak', 'cursive'].join(','),
+        fontFamily: ['Josefin Sans', 'sans-serif',].join(','),
     },
 });
 
@@ -25,11 +24,9 @@ function App() {
 
     return (
         <div className="App">
-            <UserContextProvider>
                 <MuiThemeProvider theme={theme}>
                     <Navigation/>
                 </MuiThemeProvider>
-            </UserContextProvider>
         </div>
     );
 }
