--- envolve-frontend/src/context/UserContext.jsx
@@ -1,4 +1,0 @@
-import {createContext} from 'react'
-
-export const UserStateContext = createContext({});
-export const UserDispatchContext = createContext({});
