+++ envolve-frontend/src/index.js
@@ -3,11 +3,17 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import {createStore} from 'redux'
+import allReducers from './reducers/index'
+import { Provider } from 'react-redux'
 
+const store = createStore(
+    allReducers,
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__())
 ReactDOM.render(
-
+    <Provider store={store}>
     <App />,
-
+    </Provider>,
   document.getElementById('root')
 );
