+++ envolve-frontend/src/reducers/index.js
@@ -1,0 +1,8 @@
+import UserReducer from "./UserReducer";
+import {combineReducers} from "redux";
+
+ const allReducers = combineReducers({
+    loggedUser : UserReducer,
+})
+
+export default allReducers
